Align partition on 1MiB boundary
Currently, the partition we create in 'msdos' style partitioning is
not aligned to any particular offset, and instead starts simply after
the MBR on sector 1. This creates problems with grub for example,
which doesn't have space for embedding.

To fix this and make the partitioning more standard, let's align it on
a 1MiB boundary.

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: René Nussbaumer <rn@google.com>