Check for required programs at configure time
This is not terribly helpful, especially when instance-deboostrap is
installed from packages and not from source, however it helps preventing
"obvious" missing dependencies for from-source installs.

Only tools that come from extra packages are checked, not the ones that
are installed in minimal installs  (not sure about losetup, so I've
included it as well).

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>