Switch default suite to jessie and bump min disk size
wheezy is oldstable now, so let's bump the default. Sadly, this needs
a bump in the min disk size, since on a fresh debootstrap call:

$ du -sh wheezy jessie
229M    wheezy
283M    jessie

So we're over the current 256MB limit during debootstrap run itself,
before we get a chance to cleanup deb packages, etc. I've doubled the
size since 512MB is still small, even for test instances.

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>