Add a test for exercising the scripts
To allow _some_ testing of this code without Ganeti, let's add a
simple test that exercises the create/export/import scripts.

There are a number of downsides to this script:

- it needs real root (due to losetup/kpartx/chroot), so the script
  checks for this and exits if the requirement is not met; this means
  that during normal development this will not add usefulness
- it does an actual install, so it needs network access
- although I tried to be careful by disabling cache generation and
  cleanup, it might still touch the filesystem of the host machine,
  hence it's best to run it in a throwaway VM
- it's not interactive: there's a single test, and it runs for a few
  minutes without any feedback; with automake's new test driver at
  least, there are (growing) log files in test/ which can be tail'ed

However, when plugged in and tied to (e.g.) Debian's autopkgtest
suite, it means that breakage such as
https://bugs.launchpad.net/bugs/1577346 would have been detected
automatically and much earlier.

Additionally, I had to bump the minimum automake version, since 1.9 is
no longer present in modern distributions; even wheezy (oldstable) has
1.11 as default. Tested with 1.11 and 1.15.

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>