remap the partitions after creating the filesystem
It's unclear to me why, but sometimes `mkfs` remaps the partitions and
then `blkid` cannot find them anymore.

The fun thing is that the entire script will just silently fail in
this case, because the `$VOL_ID` line will fail without an error, and
the rollback will just cleanup everything.

I had to add a lot of debugging in the script to figure out exactly
what was going on.

Not sure at all this is the right solution, but hopefully...

Closes: #13